Make computing of seed to use 64 bit to avoid sign issues in future

+++ src/gulcalc/gulcalc.cpp
@@ -321,8 +321,8 @@ damagecdfrec *d = (damagecdfrec *)rec;
 				break;
 			case rd_option::usehashedseed:
 				{
-					long s1 = (iter->group_id * 1543270363L) % 2147483648L;		// hash group_id and event_id to seed random number
-					long s2 = (d->event_id * 1943272559L) % 2147483648L;
+					unsigned long long s1 = (iter->group_id * 1543270363L) % 2147483648L;		// hash group_id and event_id to seed random number
+					unsigned long long s2 = (d->event_id * 1943272559L) % 2147483648L;
 					s1 = (s1 + s2 + rand_seed_) % 2147483648L;
 					rnd_->seedRands(s1);
 				}
